@W-12494075@: Modified rule hierarchy to properly support new rule.
