+++ sfge/src/main/java/com/salesforce/rules/AbstractPathAnomalyRule.java
@@ -1,0 +1,6 @@
+package com.salesforce.rules;
+
+/**
+ * Abstract parent class for rules that act on anomalous events that occurred during path expansion.
+ */
+public abstract class AbstractPathAnomalyRule extends AbstractPathBasedRule {}
